+++ tests/phpunit/MySQL/MySQLFieldSqlBuilderTest.php
@@ -86,7 +86,7 @@ class MySQLFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 		$argLists[] = array(
 			new FieldDefinition(
 				'fieldName',
-				FieldDefinition::TYPE_BOOLEAN,
+				FieldDefinition::TYPE_TINYINT,
 				FieldDefinition::NOT_NULL,
 				'1'
 			),